# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Product Management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/products/": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "seaech for product that matches search query in the system",
                "description": "Delete a specific product in the system",
                "operationId": "Products",
                "responses": {
                    "401": {
                        "description": "Invalid  search parameter entered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid  search parameter entered"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Save a new product into the system",
                "description": "Save a new product into the system",
                "operationId": "Products",
                "requestBody": {
                    "description": "Enter Product details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "IPhone 12"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "IPhone-12"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "IPhone 12 phone description"
                                    },
                                    "price": {
                                        "type": "decimal",
                                        "example": "4000.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Invalid  Product detials entered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid  Product Details entered "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/products/id": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product in the system",
                "description": "Get a specific product in the system",
                "operationId": "Products",
                "responses": {
                    "401": {
                        "description": "Invalid  Product ID entered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid  Product ID entered "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a specific product in the system",
                "description": "Delete a specific product in the system",
                "operationId": "Products",
                "responses": {
                    "401": {
                        "description": "Invalid  Product ID entered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid  Product ID entered "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "object",
                                            "example": "1"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/products/ID": {
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Update an existing product into the system",
                "description": "Update an existing product into the system",
                "operationId": "Products",
                "requestBody": {
                    "description": "Enter Product details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "slug",
                                    "price"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "IPhone 12"
                                    },
                                    "slug": {
                                        "type": "string",
                                        "example": "IPhone-12"
                                    },
                                    "description": {
                                        "type": "text",
                                        "example": "IPhone 12 phone description"
                                    },
                                    "price": {
                                        "type": "decimal",
                                        "example": "4000.00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Invalid  Product detials entered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid  Product Details entered "
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "product": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Product": {
                "required": [
                    "name",
                    "slug",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "maxLength": 191,
                        "example": "IPhone 12"
                    },
                    "slug": {
                        "description": "slug for the product",
                        "type": "string",
                        "readOnly": "true",
                        "example": "Iphone-12"
                    },
                    "description": {
                        "type": "text",
                        "example": "IPhone 12 description and specification of item"
                    },
                    "price": {
                        "type": "decimal",
                        "example": "241000.50"
                    },
                    "created_at": {
                        "description": "Initial creation timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Product"
                }
            },
...
        }
    },
    "security": [
        []
    ]
}